Added loader element to google actions
