+++ client/src/components/Modal/ShareModal/ShareEventModal.scss
@@ -11,7 +11,7 @@
     transform: translateX(-50%);
     border: .1rem solid $light-grey;
     border-radius: .4rem;
-    width: 9rem;
+    width: 10.5rem;
     display: flex;
     flex-direction: row;
