+++ client/styles/_variables.scss
@@ -25,6 +25,7 @@ $secondary: 'Franklin';
 // GREEN
 $primary-green: #3E5B41;
 $hover-green: #D5E2D6;
+$light-green: #1e8f29cb;
 
 // WHITE
 $white: #FFFFFF;
